refactor: RequestHandler 구현 익명함수에서 클래스로 추출
